long

+++ add_me.c
@@ -1,4 +1,5 @@
 #include <stdlib.h>
-float add_me(int x, int y) { 
+
+long add_me(int x, int y) { 
   return x + y
 }
